requirements.txt
ocrspace==2.1.1
python-dotenv==0.13.0
tweepy==3.8.0
